Rename the package of the Android app
