+++ android/app/BUCK
@@ -46,13 +46,13 @@ android_library(
 
 android_build_config(
   name = 'build_config',
-  package = 'org.jitsi.jitsi_meet_react',
+  package = 'org.jitsi.meet',
 )
 
 android_resource(
   name = 'res',
   res = 'src/main/res',
-  package = 'org.jitsi.jitsi_meet_react',
+  package = 'org.jitsi.meet',
 )
 
 android_binary(
